backend-fastapi-template/requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
sqlmodel
python-dotenv
pydantic-settings

# Auth
python-jose[cryptography]
bcrypt==4.3.0